Add support for station data / timeseries data in NetCDF
NetCDF supports storing timeseries, profiles, and trajectories using a
discrete sampling geometry (see CF-1.7 section 9)[http://cfconventions.org/Data/cf-conventions/cf-conventions-1.7/cf-conventions.html#discrete-sampling-geometries].

An instance variable can be used to add metadata; this variable is
treated in iris as a dimension coordinate. When adding the dimension
fails, it is added as an auxillary coordinate.

However, the variable does not need to be monotonic, and can be of any type.
This leads to different behaviour depending on the content of the
coordinate. And for non-numeric content, iris crashes.

This patch implements:
 * consistent loading of instance variables as auxillary coordinates,
 triggered by a valid 'cf_role' attribute
 * creating a monotonic and numeric index dimension for the series
 * fixes for dealing with string valued coordinates